@@ -1,5 +1,6 @@
 
 #include <math.h>
+#include <memory>
 #include <functional>
 #include <rrt/2dplane/2dplane.hpp>
 
@@ -7,10 +8,11 @@ using namespace Eigen;
 using namespace RRT;
 using namespace std;
 
-Tree<Vector2d>* RRT::TreeFor2dPlane(
+shared_ptr<Tree<Vector2d>> RRT::TreeFor2dPlane(
     shared_ptr<StateSpace<Eigen::Vector2d>> stateSpace, Vector2d goal,
     double step) {
-    Tree<Vector2d>* rrt = new Tree<Vector2d>(stateSpace);
+    shared_ptr<Tree<Vector2d>> rrt =
+        make_shared<Tree<Vector2d>>(stateSpace, hash, dimensions);
 
     rrt->setStepSize(step);
 
