@@ -1,4 +1,5 @@
 #include <gtest/gtest.h>
+#include <rrt/2dplane/2dplane.hpp>
 #include <rrt/2dplane/GridStateSpace.hpp>
 #include <rrt/BiRRT.hpp>
 
@@ -8,13 +9,15 @@ using namespace Eigen;
 namespace RRT {
 
 TEST(BiRRT, Instantiation) {
-    BiRRT<Vector2d> biRRT(make_shared<GridStateSpace>(50, 50, 50, 50));
+    BiRRT<Vector2d> biRRT(make_shared<GridStateSpace>(50, 50, 50, 50), hash,
+                          dimensions);
 }
 
 TEST(BiRRT, getPath) {
     Vector2d start = {1, 1}, goal = {30, 30};
 
-    BiRRT<Vector2d> biRRT(make_shared<GridStateSpace>(50, 50, 50, 50));
+    BiRRT<Vector2d> biRRT(make_shared<GridStateSpace>(50, 50, 50, 50), hash,
+                          dimensions);
     biRRT.setStartState(start);
     biRRT.setGoalState(goal);
     biRRT.setStepSize(1);
@@ -32,6 +35,23 @@ TEST(BiRRT, getPath) {
     // the path, respectively.
     EXPECT_EQ(start, path.front());
     EXPECT_EQ(goal, path.back());
+}
+
+TEST(BiRRT, multipleRuns) {
+    Vector2d start = {1, 1}, goal = {30, 30};
+
+    BiRRT<Vector2d> biRRT(make_shared<GridStateSpace>(50, 50, 50, 50), hash,
+                          dimensions);
+    biRRT.setStartState(start);
+    biRRT.setGoalState(goal);
+    biRRT.setStepSize(1);
+    biRRT.setMaxIterations(10000);
+
+    for (int i = 0; i < 50; i++) {
+        bool success = biRRT.run();
+        ASSERT_TRUE(success);
+        biRRT.reset();
+    }
 }
 
 }  // namespace RRT
